egos-sinapi-api/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Job Scheduling
apscheduler==3.10.4

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
click==8.1.7

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Logging & Observability
structlog==23.3.0
prometheus-client==0.21.0
psutil==6.1.0
